Rework transform hierarchy propagation, add PivotComponent and configurable angle units
- Rework TransformSystem descent to breadth-first depth buckets with cached generation
- Add cycle detection in hierarchy walk to terminate recursive loops
- Introduce standalone PivotComponent for anchor-based rotation and scaling
- Add configurable AngleUnit (degrees vs radians) with Lua and editor boundary conversions
- Optimize rotation math with Rotation2D caching sin/cos calculations
- Fix PositionComponent.value write-through mutation bug
- Update collision, debug draw, and primitive rendering systems to respect entity pivots
- Add LocalSize helper in General namespace and bounds-check RenderQueue capacity

+++ src/Components/RotationComponent.h
@@ -1,7 +1,9 @@
 #pragma once
 
+// Rotation about the entity's anchor, always in radians — the project's AngleUnit is converted at
+// the Lua and editor boundaries. The anchor itself lives in the optional PivotComponent.
 struct RotationComponent {
-  double value;
+  float value;
 
-  explicit RotationComponent(const double t_value = 0.0) : value(t_value) {}
+  explicit RotationComponent(const float t_value = 0.0F) : value(t_value) {}
 };
